# Areas/Admin/Models/CategoryModel.cs
namespace LMS.Areas.Admin.Models
{
    public class POSTCategoryModel
    {
        public string Name { get; set; }
    }
    public class CategoryModel
    {
        public int Id { get; set; }
        public string Name { get; set; }
    }
}